+++ src/main/java/gift/service/AuthService.java
@@ -5,21 +5,17 @@ import gift.exception.MemberException;
 import gift.model.Member;
 import gift.model.dto.MemberRequestDto;
 import gift.repository.MemberDao;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jws;
-import io.jsonwebtoken.JwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
+import gift.util.TokenProvider;
 import org.springframework.stereotype.Service;
 
 @Service
 public class AuthService {
 
-    private static final String SECRET_KEY = "Yn2kjibddFAWtnPJ2AFlL8WXmohJMCvigQggaEypa5E=";
-
+    private final TokenProvider tokenProvider;
     private final MemberDao memberDao;
 
-    public AuthService(MemberDao memberDao) {
+    public AuthService(TokenProvider tokenProvider, MemberDao memberDao) {
+        this.tokenProvider = tokenProvider;
         this.memberDao = memberDao;
     }
 
@@ -28,16 +24,7 @@ public class AuthService {
         if (!member.matchPassword(memberRequestDto.getPassword())) {
             throw new MemberException(MemberErrorCode.FAILURE_LOGIN);
         }
-        return generateToken(member);
-    }
-
-    private String generateToken(Member member) {
-        return Jwts.builder()
-            .claim("name", member.getName())
-            .claim("role", member.getRole())
-            .subject(member.getId().toString())
-            .signWith(Keys.hmacShaKeyFor(SECRET_KEY.getBytes()))
-            .compact();
+        return tokenProvider.generateToken(member);
     }
 
     public boolean validateAuthorization(String authorizationHeader) {
@@ -46,24 +33,12 @@ public class AuthService {
         }
         String type = extractType(authorizationHeader);
         String token = extractToken(authorizationHeader);
-
-        return isBearer(type) && validateToken(token);
+        return isBearer(type) && tokenProvider.validateToken(token);
     }
 
-    private boolean validateToken(String token) {
-        try {
-            Claims payload = getClaims(token);
-            return true;
-        } catch (JwtException e) {
-            return false;
-        }
-    }
-
-    public Claims getClaims(String token) throws JwtException {
-        Jws<Claims> jws = Jwts.parser().verifyWith(Keys.hmacShaKeyFor(SECRET_KEY.getBytes()))
-            .build()
-            .parseSignedClaims(token);
-        return jws.getPayload();
+    public Long getMemberId(String authorizationHeader) {
+        String token = extractToken(authorizationHeader);
+        return Long.parseLong(tokenProvider.getClaims(token).getSubject());
     }
 
     private boolean isBearer(String type) {
@@ -74,7 +49,7 @@ public class AuthService {
         return authorizationHeader.split(" ")[0];
     }
 
-    public String extractToken(String authorizationHeader) {
+    private String extractToken(String authorizationHeader) {
         return authorizationHeader.split(" ")[1];
     }
 }
