+++ src/main/java/gift/util/TokenProvider.java
@@ -1,0 +1,43 @@
+package gift.util;
+
+import gift.model.Member;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jws;
+import io.jsonwebtoken.JwtException;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.stereotype.Component;
+
+@Component
+public class TokenProvider {
+
+    private static final String SECRET_KEY = "Yn2kjibddFAWtnPJ2AFlL8WXmohJMCvigQggaEypa5E=";
+
+    public String generateToken(Member member) {
+        return Jwts.builder()
+            .claim("name", member.getName())
+            .claim("role", member.getRole())
+            .subject(member.getId().toString())
+            .signWith(Keys.hmacShaKeyFor(SECRET_KEY.getBytes()))
+            .compact();
+    }
+
+    public boolean validateToken(String token) {
+        try {
+            if (token == null) {
+                return false;
+            }
+            Claims payload = getClaims(token);
+            return true;
+        } catch (JwtException e) {
+            return false;
+        }
+    }
+
+    public Claims getClaims(String token) throws JwtException {
+        Jws<Claims> jws = Jwts.parser().verifyWith(Keys.hmacShaKeyFor(SECRET_KEY.getBytes()))
+            .build()
+            .parseSignedClaims(token);
+        return jws.getPayload();
+    }
+}
